functioning rBrief descriptors
describes image features as a 128 bit hamming code where each bit
represents a brightness comparison between a pair of points in a +/- 15
pixel window around the feature point. At each point the brightness
sample is the sum over a +/- 2 rectangle. Integral images are used to
speed the calculation of these samples.

The pairs of points are picked at random (though the same set is used
for all feature points). TODO The original ORB paper describes a learning
process to produce a better set.

To better account for rotations in the image the orientation of the
feature point is used to pick a pre-rotated version of the 128 pairs.
Note, although the original test set is within a +/- 15 pixel rectangle,
the rotated set will be within a sqrt(15^2 + 15^2) = 15 * sqrt(2)
radius.

The next step is to find and describe features in a second image and
then find matches between the features in the pair of images. That is
for each feature in image A, find its nearest neighbour in the features
in image B. Where distance between images is the hamming distance
between feature descriptors.

To accelerate the nearest neighbour matching we want to use Locality
Sensitive Hashing (see https://github.com/jeremythorne/hamming_lsh)