WIP train rBrief descriptors
as the paper makes clear, a random BRIEF descriptor derives a lot of its
power from the assumption that corners are randomly oriented. But since
we are steering the descriptor that is no longer true.

Instead we need to implement the learning algorithm described. A first
step in that is an iterator that can produce all non overlapping pairs
in a 31x31 patch with 5x5 sample windows.