implemented rBrief training algorithm
implemented algorithm from ORB 2012 paper to choose an optimal set of
binary pairs for the descriptor.

The paper isn't clear on how the orientation of the features should be
handled, so I've opted to rotate the source image so all features in the
training algorithm have the same orientation.

While the resulting patch set has a mean distance to 0.5 of about 10% of
a random set, and about half the correlation to the other members of the
set, which sounds like an improvement, matching performance with the
trained descriptors seems worse than a random set.

Perhaps this is because I'm only using a single image for training.